dagknight: address PR review comments
