+++ consensus/src/processes/dagknight/umc_voting.rs
@@ -25,9 +25,8 @@ pub trait ColoringReader {
 pub struct UmcVotingContext<'a> {
     /// The latest common chain ancestor of the zone (conflict genesis).
     pub conflict_genesis: Hash,
-    /// A member of the subgroup being evaluated; the subgroup shares one next chain
-    /// ancestor above the conflict genesis, so a single member identifies it.
-    pub subgroup_member: &'a Hash,
+    /// The next chain ancestor of the subgroup being evaluated above the conflict genesis
+    pub next_chain_ancestor: &'a Hash,
     /// The k-coloring data of the virtual GD — head of the virtual GD chain.
     pub virtual_gd: &'a GhostdagData,
     /// The rank `k` under test.
@@ -83,7 +82,7 @@ pub mod test_fixtures {
     };
     use crate::{
         model::{
-            services::reachability::MTReachabilityService,
+            services::reachability::{MTReachabilityService, ReachabilityService},
             stores::{
                 ghostdag::GhostdagData, headers::MemoryHeaderStore, reachability::MemoryReachabilityStore,
                 relations::MemoryRelationsStore,
@@ -168,7 +167,7 @@ pub mod test_fixtures {
         pub headers: Arc<MemoryHeaderStore>,
         pub reader: MemoryColoringReader,
         pub virtual_gd: Arc<GhostdagData>,
-        pub subgroup_member: Hash,
+        pub next_chain_ancestor: Hash,
         pub conflict_genesis: Hash,
         pub k: KType,
     }
@@ -226,12 +225,14 @@ pub mod test_fixtures {
                 data.virtual_gd.blue_score,
             ));
 
+            let reachability_service = MTReachabilityService::new(Arc::new(RwLock::new(reachability)));
+            let next_chain_ancestor = reachability_service.get_next_chain_ancestor(data.subgroup[0].into(), data.genesis.into());
             Self {
-                reachability: MTReachabilityService::new(Arc::new(RwLock::new(reachability))),
+                reachability: reachability_service,
                 headers,
                 reader,
                 virtual_gd,
-                subgroup_member: data.subgroup[0].into(),
+                next_chain_ancestor,
                 conflict_genesis: data.genesis.into(),
                 k: data.k,
             }
@@ -240,7 +241,7 @@ pub mod test_fixtures {
         pub fn context(&self) -> UmcVotingContext<'_> {
             UmcVotingContext {
                 conflict_genesis: self.conflict_genesis,
-                subgroup_member: &self.subgroup_member,
+                next_chain_ancestor: &self.next_chain_ancestor,
                 virtual_gd: &self.virtual_gd,
                 k: self.k,
                 coloring_reader: &self.reader,
